backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
marshmallow==3.13.0
python-dotenv==0.19.0
redis==3.5.3
pika==1.2.0
cryptography==3.4.8
gunicorn==20.1.0
alembic==1.7.1
elasticsearch==7.14.0
pytest==6.2.5
black==21.8b0
flake8==3.9.2
Flask-Migrate==3.1.0
Flask-Marshmallow==0.14.0
Marshmallow-SQLAlchemy==0.26.1
PyJWT==2.3.0
python-dateutil==2.8.2
requests==2.27.1
pytest-flask==1.2.0
Werkzeug==2.0.1 